Module imbe::consts

Constants used in the codec.

Constants

MAX_HARMONICS
Number of harmonics L when the fundamental frequency ω0 is minimum.

MAX_QUANTIZED_AMPS
Maximum number of quantized amplitudes.

MIN_HARMONICS
Number of harmonics L when the fundamental frequency ω0 is maximum.

NUM_HARMONICS
Number of discrete values that exist for the harmonics parameter.

SAMPLES_PER_FRAME
Samples per voiced/unvoiced frame

SAMPLE_RATE
Audio samples per second