Module imbe::enhance

Frame repeat/muting, spectral amplitude enhancement, and adaptive smoothing.

Structs

EnhanceErrors
Values derived from error correction decoding.

EnhancedSpectrals
Enhanced spectral amplitudes, "overbar" Ml, are derived from the decoded spectral amplitudes, "tilde" Ml.

FrameEnergy
Energy-related parameters for a voice frame.

Functions

amp_thresh
Compute the spectral amplitude threshold τM used in adaptive smoothing from the given error characteristics and previous amplitude threshold.

should_mute
Check if the current frame should be discarded and replaced with silence/comfort noise based on the given error characteristics of the current frame.

should_repeat
Check whether the current frame should be discarded and the previous repeated based on the given error characteristics of the current frame.

smooth
Smooth the given enhanced spectral amplitudes Ml and voiced/unvoiced decisions vl based on the given error characteristics, current frame energy, and spectral amplitude threshold τM for the current frame.