Function imbe::enhance::smooth

pub fn smooth(enhanced: &mut EnhancedSpectrals,
              voiced: &mut VoiceDecisions,
              errors: &EnhanceErrors,
              fen: &FrameEnergy,
              amp_thresh: f32)

Smooth the given enhanced spectral amplitudes Ml and voiced/unvoiced decisions vl based on the given error characteristics, current frame energy, and spectral amplitude threshold τM for the current frame.