Struct imbe::enhance::EnhanceErrors

pub struct EnhanceErrors {
    pub total: usize,
    pub rate: f32,
    pub golay_init: usize,
    pub hamming_init: usize,
}

Values derived from error correction decoding.

Fields
total: usize

Total number of errors corrected in the current frame, ϵT [p45].

rate: f32

Error rate tracking term, ϵR [p45].

golay_init: usize

Errors corrected in first (u0) Golay-coded chunk, ϵ0.

hamming_init: usize

Errors corrected in first (u4) Hamming-coded chunk, ϵ4.

Methods

impl EnhanceErrors

pub fn new(errors: &Errors, prev_rate: f32) -> EnhanceErrors

Create a new EnhanceErrors from the errors corrected in the current frame, ϵi, and the previous frame's ϵR value.