Struct imbe::enhance::EnhancedSpectrals

pub struct EnhancedSpectrals(_);

Enhanced spectral amplitudes, "overbar" Ml, are derived from the decoded spectral amplitudes, "tilde" Ml.

Methods

impl EnhancedSpectrals

fn new(spectrals: &Spectrals,
fen: &FrameEnergy,
params: &BaseParams)
-> EnhancedSpectrals

Create a new EnhancedSpectrals from the given base spectral amplitudes Ml and current frame energy values and parameters.

fn get(&self, l: usize) -> f32

Retrieve the enhanced spectral amplitude Ml, 1 ≤ l ≤ L.